requirements.txt
Flask
requests
Authlib
mysql-connector-python
